Add new-release job to release-pipeline.yml
A new job named `new-release` has been added to the `release-pipeline.yml` file. This job is responsible for creating a new GitHub release version and runs on `ubuntu-latest`. It is conditioned to execute only if the `new-version` job outputs a valid version and the `invoke-api-tests` job is successful. The job sets an environment variable `buildVersion` from the output of the `new-version` job.

The job consists of three steps:
1. **Checkout repository**: Uses the `actions/checkout@v4` action.
2. **Download Build Artifacts**: Uses the `actions/download-artifact@v4` action to download artifacts named `g4-hub`.
3. **Create GitHub Release & Tag**: Uses the `softprops/action-gh-release@v2` action to create a GitHub release and tag with the version specified by `buildVersion`, uploading `.zip` files and generating release notes.

The `GITHUB_TOKEN` secret is used for authentication.