Refactor resource cleanup and remove unused ESLint directive
Removed the line that disposes of the `_connection` object and
added a `console.log` statement to indicate that resources have
been cleaned up. Also, removed the ESLint directive that
disabled the `@typescript-eslint/no-unused-vars` rule for the
`StateMachineSteps` class, indicating that the class or its
variables are now being used or the directive is no longer
necessary.