Add timer management to StateMachine class
Added functionality to reset and start a timer at the beginning of the automation process in the `StateMachine` class. The timer is stopped and the display is updated with the final time elapsed after the automation process completes. This ensures accurate tracking of the time elapsed during the automation process.