Resolve #36
Enhance validation for initialValue option

Updated the validation logic for the `initialValue` option to include a check for `null` and utilize `Number.isNaN()` for improved accuracy. This change ensures that if `initialValue` is not provided, is `null`, or is NaN, it defaults to `false`; otherwise, it retains its original value.